feat(store): implement game state management with new store structure and methods
