+++ docker/media-controller/front/src/app/models/game-state.ts
@@ -1,0 +1,9 @@
+import {Player} from "./player";
+import {MachineVariables} from "../types";
+
+export interface GameState {
+  players: Player[],
+  player: number,
+  variables: MachineVariables;
+  settings: []
+}
