+++ docker/media-controller/front/src/app/store/with-game-store-computed.ts
@@ -1,0 +1,15 @@
+import {withComputed, type, signalStoreFeature} from "@ngrx/signals";
+import {computed} from "@angular/core";
+
+import { GameStoreBaseFeature } from "./game-store-base-feature";
+
+
+export function withGameStoreComputed() {
+  return signalStoreFeature(
+    type<GameStoreBaseFeature>(),
+    withComputed(({players, player}) => ({
+      currentPlayerScore: computed(() => players()[player() - 1]?.score ?? 0),
+      currentPlayerBall:  computed(() => players()[player() - 1]?.ball ?? 0),
+    }))
+  )
+}
